Avoid implicit conversions to floats in dart:ui (flutter/engine#40098)
Avoid implicit conversions to floats in dart:ui

+++ engine/src/flutter/lib/ui/painting/image_decoder_impeller.cc
@@ -141,8 +141,8 @@ std::optional<DecompressResult> ImageDecoderImpeller::DecompressTexture(
   auto decode_size = source_size;
   if (descriptor->is_compressed()) {
     decode_size = descriptor->get_scaled_dimensions(std::max(
-        static_cast<double>(target_size.width()) / source_size.width(),
-        static_cast<double>(target_size.height()) / source_size.height()));
+        static_cast<float>(target_size.width()) / source_size.width(),
+        static_cast<float>(target_size.height()) / source_size.height()));
   }
 
   //----------------------------------------------------------------------------
